Implementación de policy para libros e implementación de agregado y eliminación de ejemplares por libro

+++ app/Http/Controllers/LibroController.php
@@ -3,11 +3,14 @@
 namespace App\Http\Controllers;
 
 use App\Models\Libro;
+use App\Models\Ejemplar;
 use Illuminate\Http\Request;
 use App\Models\Categoria;
+use Illuminate\Database\Eloquent\Builder;
 
 class LibroController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      *
@@ -15,7 +18,12 @@ class LibroController extends Controller
      */
     public function index()
     {
-        $libros = Libro::all();
+        $libros = Libro::with('ejemplares')->withCount([
+            'ejemplares',
+            'ejemplares as ejemplares_en_prestamo' => function(Builder $query) {
+                $query->where('en_prestamo', true);
+            },
+        ])->get();
         return view('libros/libroIndex', compact('libros'));
     }
 
@@ -26,6 +34,14 @@ class LibroController extends Controller
      */
     public function create()
     {
+        if(!\Auth::user()->can('create', Libro::class))
+        {
+            return redirect()->route('libro.index')->with([
+                'mensaje-alerta' => 'No cuenta con los permisos necesarios.',
+                'titulo-alerta' => 'Acceso denegado!',
+                'tipo-alerta' => 'alert-danger',
+            ]);
+        }
         $categorias = Categoria::All();
         return view('libros/libroForm', compact('categorias'));
     }
@@ -38,6 +54,15 @@ class LibroController extends Controller
      */
     public function store(Request $request)
     {
+        if(!\Auth::user()->can('create', Libro::class))
+        {
+            return redirect()->route('libro.index')->with([
+                'mensaje-alerta' => 'No cuenta con los permisos necesarios.',
+                'titulo-alerta' => 'Acceso denegado!',
+                'tipo-alerta' => 'alert-danger',
+            ]);
+        }
+
         $request->validate([
             'isbn' => 'required|unique:libros|digits_between:10,13',
             'nombre' => 'required|max:100',
@@ -57,7 +82,7 @@ class LibroController extends Controller
         Libro::create($libro);
 
         return redirect()->route('libro.index')->with([
-            'mensaje-alerta' => 'Libro agregado exitosamente.',
+            'mensaje-alerta' => 'Libro ' . $libro->nombre . ' agregado exitosamente.',
             'titulo-alerta' => 'Acción exitosa!',
             'tipo-alerta' => 'alert-success',
         ]);
@@ -71,11 +96,14 @@ class LibroController extends Controller
      */
     public function show(Libro $libro)
     {
-        return redirect()->route('libro.index')->with([
-            'mensaje-alerta' => 'La página a la que intentaste acceder no existe',
-            'titulo-alerta' => 'Oops...',
-            'tipo-alerta' => 'alert-info',
-        ]);
+        if(!\Auth::user()->can('view', Libro::class))
+        {
+            return redirect()->route('libro.index')->with([
+                'mensaje-alerta' => 'La página a la que intentaste acceder no existe',
+                'titulo-alerta' => 'Oops...',
+                'tipo-alerta' => 'alert-info',
+            ]);
+        }
     }
 
     /**
@@ -86,6 +114,14 @@ class LibroController extends Controller
      */
     public function edit(Libro $libro)
     {
+        if(!\Auth::user()->can('update', Libro::class))
+        {
+            return redirect()->route('libro.index')->with([
+                'mensaje-alerta' => 'No cuenta con los permisos necesarios.',
+                'titulo-alerta' => 'Acceso denegado!',
+                'tipo-alerta' => 'alert-danger',
+            ]);
+        }
         $categorias = Categoria::All();
         return view('libros/libroForm', compact('libro', 'categorias'));
     }
@@ -99,6 +135,14 @@ class LibroController extends Controller
      */
     public function update(Request $request, Libro $libro)
     {
+        if(!\Auth::user()->can('update', Libro::class))
+        {
+            return redirect()->route('libro.index')->with([
+                'mensaje-alerta' => 'No cuenta con los permisos necesarios.',
+                'titulo-alerta' => 'Acceso denegado!',
+                'tipo-alerta' => 'alert-danger',
+            ]);
+        }
         $request->validate([
             'nombre' => 'required|max:100',
             'autor' => 'nullable|max:255',
@@ -119,7 +163,7 @@ class LibroController extends Controller
         $libro->save();
 
         return redirect()->route('libro.index')->with([
-            'mensaje-alerta' => 'Libro editado exitosamente.',
+            'mensaje-alerta' => 'Libro ' . $libro->isbn . ' editado exitosamente.',
             'titulo-alerta' => 'Acción exitosa!',
             'tipo-alerta' => 'alert-success',
         ]);
@@ -133,9 +177,80 @@ class LibroController extends Controller
      */
     public function destroy(Libro $libro)
     {
+        if(!\Auth::user()->can('delete', Libro::class))
+        {
+            return redirect()->route('libro.index')->with([
+                'mensaje-alerta' => 'No cuenta con los permisos necesarios.',
+                'titulo-alerta' => 'Acceso denegado!',
+                'tipo-alerta' => 'alert-danger',
+            ]);
+        }
+
         $libro->delete();
         return redirect()->route('libro.index')->with([
             'mensaje-alerta' => 'Libro borrado exitosamente.',
+            'titulo-alerta' => 'Acción exitosa!',
+            'tipo-alerta' => 'alert-success',
+        ]);
+    }
+
+    public function agregarEjemplar(Libro $libro)
+    {
+        if(!\Auth::user()->can('update', Libro::class))
+        {
+            return redirect()->route('libro.index')->with([
+                'mensaje-alerta' => 'No cuenta con los permisos necesarios.',
+                'titulo-alerta' => 'Acceso denegado!',
+                'tipo-alerta' => 'alert-danger',
+            ]);
+        }
+
+        $numero = Ejemplar::where('isbn', $libro->isbn)->max('numero');
+
+        Ejemplar::create([
+            'isbn' => $libro->isbn,
+            'numero' => $numero + 1,
+        ]);
+
+        return redirect()->route('libro.index')->with([
+            'mensaje-alerta' => 'Ejemplar de '. $libro->nombre .' agregado exitosamente.',
+            'titulo-alerta' => 'Acción exitosa!',
+            'tipo-alerta' => 'alert-success',
+        ]);
+    }
+
+    public function eliminarEjemplar(Libro $libro)
+    {
+        if(!\Auth::user()->can('update', Libro::class))
+        {
+            return redirect()->route('libro.index')->with([
+                'mensaje-alerta' => 'No cuenta con los permisos necesarios.',
+                'titulo-alerta' => 'Acceso denegado!',
+                'tipo-alerta' => 'alert-danger',
+            ]);
+        }
+
+        $numero = Ejemplar::where('isbn', $libro->isbn)->max('numero');
+
+        if($numero == 0)
+            return redirect()->route('libro.index')->with([
+                'mensaje-alerta' => 'Este libro no cuenta con ejemplares.',
+                'titulo-alerta' => 'Oops...',
+                'tipo-alerta' => 'alert-info',
+            ]);
+
+        $Ejemplar = Ejemplar::where('isbn', $libro->isbn)->where('numero', $numero);
+        $Ejemplar->delete();
+
+        if($numero == 2)
+            return redirect()->route('libro.index')->with([
+                'mensaje-alerta' => 'Ejemplar de ' . $libro->nombre . ' eliminado exitosamente. Unicamente queda un ejemplar!',
+                'titulo-alerta' => 'Advertencia!',
+                'tipo-alerta' => 'alert-warning',
+            ]);
+
+        return redirect()->route('libro.index')->with([
+            'mensaje-alerta' => 'Ejemplar de ' . $libro->nombre . ' eliminado exitosamente.',
             'titulo-alerta' => 'Acción exitosa!',
             'tipo-alerta' => 'alert-success',
         ]);
